Milestone checkin

+++ src/com/mailbrew/commands/DeleteAccountCommand.as
@@ -7,14 +7,20 @@ package com.mailbrew.commands
 	import com.mailbrew.database.DatabaseResponder;
 	import com.mailbrew.events.DeleteAccountEvent;
 	import com.mailbrew.events.PopulateAccountListEvent;
+	import com.mailbrew.events.UpdateAppIconEvent;
 	import com.mailbrew.model.ModelLocator;
+	
+	import flash.display.NativeMenu;
 	
 	public class DeleteAccountCommand
 		implements ICommand
 	{
 		public function execute(e:CairngormEvent):void
 		{
-			var ml:ModelLocator = ModelLocator.getInstance();
+			if (ModelLocator.getInstance().checkEmailLock)
+			{
+				return;
+			}
 			var dae:DeleteAccountEvent = DeleteAccountEvent(e);
 			this.deleteMessages(dae.accountId);
 		}
@@ -39,9 +45,26 @@ package com.mailbrew.commands
 				responder.removeEventListener(DatabaseEvent.RESULT_EVENT, listener);
 				ModelLocator.getInstance().accountInfo = null;
 				new PopulateAccountListEvent().dispatch();
+				removeIconMenu(accountId);
 			};
 			responder.addEventListener(DatabaseEvent.RESULT_EVENT, listener);
 			ModelLocator.getInstance().db.deleteAccountById(responder, accountId);
+		}
+		
+		private function removeIconMenu(accountId:Number):void
+		{
+			var menu:NativeMenu = ModelLocator.getInstance().purr.getMenu();
+			if (menu.getItemByName(String(accountId)) != null)
+			{
+				menu.removeItem(menu.getItemByName(String(accountId)));
+			}
+			this.updateAppIcon();
+		}
+		
+		private function updateAppIcon():void
+		{
+			var uaie:UpdateAppIconEvent = new UpdateAppIconEvent();
+			uaie.dispatch();
 		}
 	}
 }
